cli for curler
